Change absolute paths to relative paths (#3)
path.module returns an absolute path. This causes terraform to mark
the lambda function as changing whenever a plan is performed at a
different path. This happens often when different users perform a
plan or apply. In Terraform v0.12 path.module returns a relative
path. To work around this issue in version 0.11 we are hard coding
a relative path that matches the value of path.module in v0.12.

In addition, variables were added for testing & portability, and
the documentation was updated to reflect these changes.

Closes #1